+++ tools/hulkz-viewer/src/app/state.rs
@@ -5,176 +5,18 @@ use std::{
     time::{Duration, Instant},
 };
 
-use eframe::egui;
 use egui_dock::DockState;
-use hulkz_stream::{BackendStats, PlaneKind, SourceStats};
+use hulkz_stream::{BackendStats, SourceStats};
 use serde::{Deserialize, Serialize};
 use tokio::{runtime::Runtime, sync::mpsc};
 use tokio_util::sync::CancellationToken;
 
-use crate::model::{
-    DiscoveredParameter, DiscoveredPublisher, DiscoveredSession, DisplayedRecord,
-    ParameterReference, SourceBindingRequest, StreamId, ViewerConfig, WorkerCommand,
-    WorkerEventEnvelope,
+use crate::protocol::{
+    DiscoveredParameter, DiscoveredPublisher, DiscoveredSession, DisplayedRecord, StreamId,
+    ViewerConfig, WorkerCommand, WorkerEventEnvelope,
 };
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub(super) enum NamespaceSelection {
-    #[default]
-    FollowDefault,
-    Override(String),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(super) struct TextPanelTab {
-    pub(super) id: StreamId,
-    #[serde(default)]
-    pub(super) namespace_selection: NamespaceSelection,
-    pub(super) source_expression: String,
-}
-
-impl TextPanelTab {
-    pub(super) fn new(id: StreamId, source_expression: String) -> Self {
-        Self {
-            id,
-            namespace_selection: NamespaceSelection::FollowDefault,
-            source_expression,
-        }
-    }
-
-    pub(super) fn follows_default_namespace(&self) -> bool {
-        matches!(self.namespace_selection, NamespaceSelection::FollowDefault)
-    }
-
-    pub(super) fn set_namespace_override_enabled(
-        &mut self,
-        enabled: bool,
-        default_namespace: &str,
-    ) {
-        if enabled {
-            let default_namespace = default_namespace.trim().to_string();
-            self.namespace_selection = NamespaceSelection::Override(default_namespace);
-        } else {
-            self.namespace_selection = NamespaceSelection::FollowDefault;
-        }
-    }
-
-    pub(super) fn namespace_override_text_mut(&mut self) -> Option<&mut String> {
-        match &mut self.namespace_selection {
-            NamespaceSelection::FollowDefault => None,
-            NamespaceSelection::Override(value) => Some(value),
-        }
-    }
-
-    pub(super) fn effective_namespace(&self, default_namespace: &str) -> Option<String> {
-        let raw = match &self.namespace_selection {
-            NamespaceSelection::FollowDefault => default_namespace,
-            NamespaceSelection::Override(value) => value,
-        };
-        let namespace = raw.trim();
-        if namespace.is_empty() {
-            None
-        } else {
-            Some(namespace.to_string())
-        }
-    }
-
-    pub(super) fn binding_request(&self, default_namespace: &str) -> Option<SourceBindingRequest> {
-        let namespace = self.effective_namespace(default_namespace)?;
-        let path_expression = self.source_expression.trim().to_string();
-        if path_expression.is_empty() {
-            return None;
-        }
-        Some(SourceBindingRequest {
-            namespace,
-            plane: PlaneKind::View,
-            path_expression,
-        })
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(super) struct ParameterPanelTab {
-    pub(super) id: u64,
-    #[serde(default)]
-    pub(super) namespace_selection: NamespaceSelection,
-    pub(super) node_input: String,
-    pub(super) path_input: String,
-    pub(super) editor_text: String,
-    #[serde(skip)]
-    pub(super) selected_parameter_reference: Option<ParameterReference>,
-    #[serde(default)]
-    pub(super) status: Option<ParameterPanelStatus>,
-}
-
-impl ParameterPanelTab {
-    pub(super) fn new(id: u64) -> Self {
-        Self {
-            id,
-            namespace_selection: NamespaceSelection::FollowDefault,
-            node_input: String::new(),
-            path_input: String::new(),
-            editor_text: String::new(),
-            selected_parameter_reference: None,
-            status: None,
-        }
-    }
-
-    pub(super) fn effective_namespace(&self, default_namespace: &str) -> Option<String> {
-        let raw = match &self.namespace_selection {
-            NamespaceSelection::FollowDefault => default_namespace,
-            NamespaceSelection::Override(value) => value,
-        };
-        let namespace = raw.trim();
-        if namespace.is_empty() {
-            None
-        } else {
-            Some(namespace.to_string())
-        }
-    }
-
-    pub(super) fn set_selected_parameter_reference(&mut self, target: ParameterReference) {
-        self.node_input = target.node.clone();
-        self.path_input = target.path_expression.clone();
-        self.selected_parameter_reference = Some(target);
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(super) struct ParameterPanelStatus {
-    pub(super) success: bool,
-    pub(super) message: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(super) enum ViewerTab {
-    #[serde(alias = "TextStreamPanel")]
-    Text(TextPanelTab),
-    Parameters(ParameterPanelTab),
-}
-
-impl ViewerTab {
-    pub(super) fn title_label(&self) -> &'static str {
-        match self {
-            ViewerTab::Text(_) => "Text",
-            ViewerTab::Parameters(_) => "Parameters",
-        }
-    }
-
-    pub(super) fn dock_id(&self) -> egui::Id {
-        match self {
-            ViewerTab::Text(stream) => egui::Id::new(("viewer_tab_text", stream.id)),
-            ViewerTab::Parameters(panel) => egui::Id::new(("viewer_tab_parameters", panel.id)),
-        }
-    }
-
-    pub(super) fn is_closeable(&self, text_panel_count: usize) -> bool {
-        match self {
-            ViewerTab::Text(_) => text_panel_count > 1,
-            ViewerTab::Parameters(_) => true,
-        }
-    }
-}
+use super::workspace_panel::WorkspacePanel;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub(super) struct PersistedUiState {
@@ -295,9 +137,9 @@ pub(super) struct TimelineState {
 
 #[derive(Debug)]
 pub(super) struct WorkspaceState {
-    pub(super) dock_state: DockState<ViewerTab>,
+    pub(super) dock_state: DockState<WorkspacePanel>,
     pub(super) stream_states: BTreeMap<StreamId, StreamRuntimeState>,
-    pub(super) binding_cache: BTreeMap<StreamId, Option<SourceBindingRequest>>,
+    pub(super) binding_cache: BTreeMap<StreamId, Option<crate::protocol::SourceBindingRequest>>,
     pub(super) next_stream_id: StreamId,
     pub(super) next_parameter_panel_id: u64,
 }
@@ -331,7 +173,7 @@ pub(super) struct UiState {
 pub(super) const MIN_TIMELINE_SPAN: Duration = Duration::from_millis(50);
 pub(super) const DEFAULT_TIMELINE_LANE_HEIGHT_PX: f32 = 22.0;
 
-pub(crate) struct ViewerApp {
+pub struct ViewerApp {
     pub(super) config: ViewerConfig,
     pub(super) shell: ShellState,
     pub(super) discovery: DiscoveryState,
@@ -354,46 +196,7 @@ const fn default_true() -> bool {
 
 #[cfg(test)]
 mod tests {
-    use super::{NamespaceSelection, PersistedUiState, TextPanelTab};
-
-    #[test]
-    fn text_panel_follow_default_namespace_tracks_input() {
-        let panel = TextPanelTab::new(1, "odometry".to_string());
-        let first = panel
-            .binding_request("robot-a")
-            .expect("valid binding with default namespace");
-        let second = panel
-            .binding_request("robot-b")
-            .expect("valid binding with updated namespace");
-
-        assert_eq!(first.namespace, "robot-a");
-        assert_eq!(second.namespace, "robot-b");
-    }
-
-    #[test]
-    fn text_panel_override_namespace_ignores_default_changes() {
-        let mut panel = TextPanelTab::new(2, "odometry".to_string());
-        panel.namespace_selection = NamespaceSelection::Override("robot-x".to_string());
-
-        let first = panel
-            .binding_request("robot-a")
-            .expect("valid binding with namespace override");
-        let second = panel
-            .binding_request("robot-b")
-            .expect("valid binding with namespace override");
-
-        assert_eq!(first.namespace, "robot-x");
-        assert_eq!(second.namespace, "robot-x");
-    }
-
-    #[test]
-    fn text_panel_binding_requires_namespace_and_path() {
-        let mut panel = TextPanelTab::new(3, "".to_string());
-        assert!(panel.binding_request("demo").is_none());
-
-        panel.source_expression = "odometry".to_string();
-        assert!(panel.binding_request("").is_none());
-    }
+    use super::PersistedUiState;
 
     #[test]
     fn persisted_ui_defaults_allow_missing_fields() {
