+++ tools/hulkz-viewer/src/app/workspace_panels/parameters.rs
@@ -1,19 +1,71 @@
-use eframe::egui::{self, Color32};
 use hulk_widgets::CompletionEdit;
+use serde::{Deserialize, Serialize};
 
-use crate::model::WorkerCommand;
+use crate::app::panel_prelude::{egui, Panel, PanelContext, UiIntent};
+use crate::protocol::ParameterReference;
 
-use super::{
-    super::state::{NamespaceSelection, ParameterPanelStatus, ParameterPanelTab},
-    Panel, ViewerApp,
-};
+use super::shared::NamespaceSelection;
 
-pub(super) struct ParametersPanel;
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ParametersWorkspacePanelState {
+    pub(crate) id: u64,
+    #[serde(default)]
+    pub(crate) namespace_selection: NamespaceSelection,
+    pub(crate) node_input: String,
+    pub(crate) path_input: String,
+    pub(crate) editor_text: String,
+    #[serde(skip)]
+    pub(crate) selected_parameter_reference: Option<ParameterReference>,
+    #[serde(default)]
+    pub(crate) status: Option<ParametersPanelStatus>,
+}
 
-impl Panel for ParametersPanel {
-    type State = ParameterPanelTab;
+impl ParametersWorkspacePanelState {
+    pub fn new(id: u64) -> Self {
+        Self {
+            id,
+            namespace_selection: NamespaceSelection::FollowDefault,
+            node_input: String::new(),
+            path_input: String::new(),
+            editor_text: String::new(),
+            selected_parameter_reference: None,
+            status: None,
+        }
+    }
 
-    fn draw(app: &mut ViewerApp, ui: &mut egui::Ui, panel: &mut Self::State) {
+    pub fn effective_namespace(&self, default_namespace: &str) -> Option<String> {
+        let raw = match &self.namespace_selection {
+            NamespaceSelection::FollowDefault => default_namespace,
+            NamespaceSelection::Override(value) => value,
+        };
+        let namespace = raw.trim();
+        if namespace.is_empty() {
+            None
+        } else {
+            Some(namespace.to_string())
+        }
+    }
+
+    pub fn set_selected_parameter_reference(&mut self, target: ParameterReference) {
+        self.node_input = target.node.clone();
+        self.path_input = target.path_expression.clone();
+        self.selected_parameter_reference = Some(target);
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ParametersPanelStatus {
+    pub(crate) success: bool,
+    pub(crate) message: String,
+}
+
+pub struct ParametersWorkspacePane;
+
+impl Panel for ParametersWorkspacePane {
+    type State = ParametersWorkspacePanelState;
+
+    fn draw(ctx: &mut PanelContext<'_>, ui: &mut egui::Ui, panel: &mut Self::State) {
+        let default_namespace = ctx.app().ui.default_namespace.clone();
         ui.horizontal(|ui| {
             let mut override_enabled =
                 !matches!(panel.namespace_selection, NamespaceSelection::FollowDefault);
@@ -22,7 +74,7 @@ impl Panel for ParametersPanel {
                 .changed()
             {
                 panel.namespace_selection = if override_enabled {
-                    NamespaceSelection::Override(app.ui.default_namespace.trim().to_string())
+                    NamespaceSelection::Override(default_namespace.trim().to_string())
                 } else {
                     NamespaceSelection::FollowDefault
                 };
@@ -34,7 +86,7 @@ impl Panel for ParametersPanel {
             }
         });
 
-        let Some(namespace) = panel.effective_namespace(&app.ui.default_namespace) else {
+        let Some(namespace) = panel.effective_namespace(&default_namespace) else {
             ui.add_space(6.0);
             ui.label("Set a default namespace first.");
             return;
@@ -44,7 +96,7 @@ impl Panel for ParametersPanel {
         let mut trigger_auto_load = false;
         ui.horizontal(|ui| {
             ui.label("Node");
-            let node_candidates = app.parameter_node_candidates(panel);
+            let node_candidates = ctx.app().parameter_node_candidates(panel);
             let node_edit = ui.add(
                 CompletionEdit::new(
                     ui.id().with(("parameter_node", panel.id)),
@@ -66,7 +118,7 @@ impl Panel for ParametersPanel {
 
             ui.label("Path")
                 .on_hover_text("DSL: gain | /global/param | ~node/private_param");
-            let path_candidates = app.parameter_path_candidates(panel);
+            let path_candidates = ctx.app().parameter_path_candidates(panel);
             let path_edit = ui.add(
                 CompletionEdit::new(
                     ui.id().with(("parameter_path", panel.id)),
@@ -88,30 +140,43 @@ impl Panel for ParametersPanel {
         });
 
         if trigger_auto_load {
-            if let Ok(target) = app.parameter_reference_from_inputs(panel) {
+            let target_result = ctx.app().parameter_reference_from_inputs(panel);
+            if let Ok(target) = target_result {
                 let should_read = panel.selected_parameter_reference.as_ref() != Some(&target);
                 panel.set_selected_parameter_reference(target.clone());
                 if should_read {
                     panel.status = None;
-                    app.send_command(WorkerCommand::ReadParameter(target));
+                    ctx.emit(UiIntent::ReadParameter {
+                        panel_id: panel.id,
+                        target,
+                    });
                 }
             } else if panel.node_input.trim().is_empty() && panel.path_input.trim().is_empty() {
                 panel.selected_parameter_reference = None;
             }
         }
 
+        let discovered_parameter_count = ctx
+            .app()
+            .discovery
+            .parameters
+            .iter()
+            .filter(|parameter| parameter.namespace == namespace_filter)
+            .count();
         ui.horizontal(|ui| {
             if ui.button("Apply").clicked() {
-                match app.parameter_reference_from_inputs(panel) {
+                let apply_target = ctx.app().parameter_reference_from_inputs(panel);
+                match apply_target {
                     Ok(target) => {
                         panel.set_selected_parameter_reference(target.clone());
-                        app.send_command(WorkerCommand::SetParameter {
+                        ctx.emit(UiIntent::ApplyParameter {
+                            panel_id: panel.id,
                             target,
                             value_json: panel.editor_text.clone(),
                         });
                     }
                     Err(message) => {
-                        panel.status = Some(ParameterPanelStatus {
+                        panel.status = Some(ParametersPanelStatus {
                             success: false,
                             message,
                         });
@@ -120,12 +185,7 @@ impl Panel for ParametersPanel {
             }
             ui.separator();
             ui.small(format!(
-                "{} parameter(s) discovered",
-                app.discovery
-                    .parameters
-                    .iter()
-                    .filter(|parameter| parameter.namespace == namespace_filter)
-                    .count()
+                "{discovered_parameter_count} parameter(s) discovered"
             ));
         });
 
@@ -143,9 +203,9 @@ impl Panel for ParametersPanel {
         );
         if let Some(status) = &panel.status {
             if status.success {
-                ui.colored_label(Color32::GREEN, &status.message);
+                ui.colored_label(egui::Color32::GREEN, &status.message);
             } else {
-                ui.colored_label(Color32::RED, &status.message);
+                ui.colored_label(egui::Color32::RED, &status.message);
             }
         }
     }
