+++ tools/hulkz-viewer/src/app/workspace_panels/shared.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
+pub enum NamespaceSelection {
+    #[default]
+    FollowDefault,
+    Override(String),
+}
