+++ tools/hulkz-viewer/src/main.rs
@@ -1,16 +1,10 @@
-mod app;
-mod model;
-mod timeline_canvas;
-mod worker;
-
-use app::{ViewerApp, ViewerStartupOverrides};
 use clap::Parser;
 use color_eyre::{
     eyre::{eyre, WrapErr as _},
     Result,
 };
+use hulkz_viewer::{setup_logging, ViewerApp, ViewerStartupOverrides};
 use std::path::PathBuf;
-use tracing_subscriber::{layer::SubscriberExt as _, util::SubscriberInitExt as _, EnvFilter};
 
 #[derive(Debug, Parser)]
 #[command(
@@ -28,27 +22,6 @@ struct Cli {
     /// Use persistent storage at the provided path instead of session temp storage
     #[arg(long)]
     storage_path: Option<PathBuf>,
-}
-
-fn setup_logging() -> Result<()> {
-    let filter = EnvFilter::try_from_default_env()
-        .unwrap_or_else(|_| EnvFilter::new("hulkz_viewer=info,hulkz_stream=info,warn"));
-
-    let layer = tracing_subscriber::fmt::layer()
-        .with_target(true)
-        .with_thread_ids(false)
-        .with_level(true)
-        .with_file(false)
-        .with_line_number(false)
-        .compact();
-
-    tracing_subscriber::registry()
-        .with(filter)
-        .with(layer)
-        .try_init()
-        .wrap_err("failed to initialize tracing subscriber")?;
-
-    Ok(())
 }
 
 fn main() -> Result<()> {
